+++ Changelog.md
@@ -1,7 +1,13 @@
 Changelog
 =========
 
-### Version 0.7 (2026-03-02)
+
+### Version Sharp (2026-03-03)
+* Use Sharp for image optimization
+* Improve error handling and logging
+
+### Version squoosh (2026-03-02)
 * Create Dockerfile
 * Create Github repository
 * Create an image in the Docker hub
+* Use Squoosh for image optimization